More examples and a brain teaser
I can't quite figure out the right incantation to be able to pass a
reference back outside the async future. I think it might be impossible,
but I have yet to convince myself that it is. I'm going to switch tack
later and attempt to pass a reference into the closure.

I remember reading that there were issues stabilizing async due to
references and await points. Shipping references in and out of the
future kinda feels like I might be hitting similar issues (in spirit at
least). I'll go back and attempt to read up on some of those details in
the hopes I'll either figure it out, or confirm that its impossible. Or
perhaps one of those "possible, *if* it were implemented inside the
compiler" type cases.